Added a "top_ancestor_id" field to Vote and Score for faster filtering within a community or panel.
Because these models had GenericForeignKeys, we were unable to filter on the fields of the related objects (e.g. the top_ancestor field, which is the top ancestor in the hierarchy, usually a community or panel).  Since we use the top ancestor for permissions and often want to gather recommendations across the visible communities, this filter is important.

We did opt to make it a PositiveIntegerField, like the object_id, because votes may exist for old communities, and we didn't want to create too hard of a dependency between the apps.